+++ CHANGELOG.md
@@ -6,7 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.5.0] - 2025-03-19
+## [0.5.1] - 2025-09-23
+
+### Changed
+
+- Limit now works differently for fluids and energies
+- Method for validating items now just `get` for advanced digitizer
+
+## [0.5.0] - 2025-09-22
 
 ### Added